requirements.txt
fastapi
uvicorn
torch==2.4.1+cpu
torchvision==0.19.1+cpu
pillow
numpy
opencv-python-headless
mediapipe
python-multipart